refactor: api/spell 폴더 하위로 이동
