+++ src/queries/useSpellCheck.ts
@@ -4,11 +4,11 @@ import { type AxiosError, type AxiosResponse } from 'axios';
 import { SPELLCHECK_STATE_MESSAGE } from '@constants/spellCheck';
 import { useToastStore } from '@stores/toast';
 
-import spellCheckApi from '../api/spellCheck';
+import spellCheckApi from '../api/spell/spellCheck';
 import {
   type SpellCheckParams,
   type SpellCheckResponse,
-} from '../types/spellCheck';
+} from '../api/spell/types';
 
 export const useSpellCheck = (
   options?: UseMutationOptions<
